Move squad functionality
A couple of things here:

ColorGraph is now a singleton, and has a couple of new methods to expose
data. StrategyManager was updated accordingly.
Yellow is no longer a color we care about.

In CombatCommander, there is some test logic to move all combat units to
a location (this is randomly generated every time squads are updated, so
they sort of just wander around spastically). This has made it virtually
impossible for the bot to win, unless you get SUPER lucky with regions.