Integration of GoalAdvisor and troop movement
Edited GoalAdvisor so it would compile and modified CombatCommander and
GameCommander to integrate it with unit movement commands.
Also made it a singleton, to be more in line with the existing advisors
and managers.
Finally, I moved the source files into the source directory, since they
were not #include-ing properly in the top-level directory.