Scout Manager can scan but cant build comsats??
Scout manager has the ability to scan the goalNode now if it has enough
energy and has a comsat built. However right now, it can't build a
comsat because the SCVs are always in the way. So it skips it
completely. I changed the building padding to 2 because in some cases
the starting position was allowing for the comsat to be built, but a
barracks was build directly to the right of the command center, blocking
the addon building. So this works *sometimes*: when the mineral fields
are to the right of the starting command center.