ScoutManager changed/updated
Note: this code hasn't been tested yet because my debugger won't work,
I'm missing pdb files for loading the symbols for debugging with
Starcraft, and I'm not sure how to fix this. Hopefully when we meet
again we can look at it, test it, debug it, and verify that it works as
it should. When I run it without debugging, it looks like its setting
all workers' jobs to WorkerData::Scout, but not sure why. When I can
debug, I will probably be able to figure this out.

Scrapped trying to get comsats to work, its just too complicated to try
and get the game to build it. Instead, adapted the current code to use
vultures. At most two vultures are used for scouting purposes at once.
They act slightly more aggressive than SCV scouts. Also added updating
of the ColorGraph based on what the scouts encounter.