Scout Manager update
Now that my deubgger works I can fix some of the issues. Also added a
draw for debug info in scout manager. Tried working on integrating scout
information into working manager, like it was supposed to be all along
according to the comments I saw. Making it realize the single worker
scout its using exists was hard to get working, but it works. Now
there's a problem in the build order manager, it's crashing on accessing
a vulture to build somewhere? I can't decipher it. Good place to stop