various improvements to build order
1. Added logic to make sure add ons are not added to the build order
unless there is an available building already in the build order to add
on to.
2. Changed excessive buildings to be based on how many units in the goal
will be produced by that building total. i.e. if we are trying to build
5 tanks and 2 vultures that is 7 units total from the factory.
3. Made the build order pop off a random action rather than just pop the
first action from the list.
4. Changed strategy manager to look at all the units that are complete
plus manually find the units that are being trained. The
BWAPI::Broodwar->self()->allUnitCount function was returning different
information from completed + training units, which was what I expected.
5. Building manager updated so that it handles the special case of
building add ons. Allowing the build manager to handle it made it so
that it never fails.